[14.0][FIX] product_configurator_mrp (Operation Steps)
This fixes an issue when a new variant is created, it creates the BoM with operations and steps, since it's copying the steps, the product on the steps gets set with the product template and not the new variant. So, when an MO is created for the variant, the work order steps don't show, because they don't apply to the variant.
The change is to simply set the product on the step to the new variant.